[Detection] Camera: Refactor code
